requirements.txt
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
boto3==1.9.43
botocore==1.12.253
Brlapi==0.8.2
brotlipy==0.7.0
bs4==0.0.1
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.6
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==7.1.2
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
csscompressor==0.9.5
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.1
distro==1.5.0
distro-info==1.0
dj-database-url==0.5.0
Django==4.1
django-appconf==1.0.4
django-compressor==2.4.1
django-cors-headers==3.8.0
django-debug-toolbar==3.2.2
django-dotenv==1.4.2
django-elasticsearch-dsl==7.2.2
django-environ==0.9.0
django-filter==22.1
django-recaptcha==2.0.6
django-sass-processor==1.0.1
django-sslserver==0.22
django-storages==1.7.1
django-suit @ https://github.com/darklow/django-suit/tarball/v2
django-templated-mail==1.1.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docker==4.4.4
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
drf-yasg==1.21.3
duplicity==0.8.17
elasticsearch==7.10.1
elasticsearch-dsl==7.4.0
environ==1.0
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
gpg===1.14.0-unknown
gunicorn==20.1.0
gyp==0.1
httplib2==0.18.1
idna==2.10
importlib-metadata==1.6.0
inflection==0.5.1
install==1.3.4
itypes==1.2.0
jeepney==0.6.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
keyring==22.2.0
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.5
libsass==0.21.0
lockfile==0.12.2
louis==3.16.0
lxml==4.8.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.9
oauthlib==3.1.0
olefile==0.46
openapi==1.1.0
packaging==21.3
paramiko==2.7.2
pexpect==4.8.0
phonenumbers==8.12.40
Pillow==8.1.2
protobuf==3.12.4
psycopg2-binary==2.9.1
pycairo==1.16.2
pycparser==2.20
pycups==2.0.1
PyGObject==3.38.0
PyJWT==2.1.0
pymacaroons==0.13.0
PyNaCl==1.4.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt===2.2.0-ubuntu0.21.04.1
python-dateutil==2.8.1
python-debian==0.1.39
python-dotenv==0.19.0
python3-openid==3.2.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.3.1
rcssmin==1.0.6
reportlab==3.5.66
requests==2.25.1
requests-oauthlib==1.3.0
rjsmin==1.1.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.1.13
SecretStorage==3.3.1
simplejson==3.17.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
sqlparse==0.4.1
systemd-python==234
texttable==1.6.4
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.25.11
usb-creator==0.3.7
virtualenv==20.4.0+ds
wadllib==1.3.5
websocket-client==0.59.0
whitenoise==3.3.1
xdg==5
xkit==0.0.0
xlwt==1.3.0
zipp==1.0.0
